extract io utils
