+++ src/moonlink/src/storage/iceberg.rs
@@ -5,6 +5,7 @@ mod iceberg_table_loader;
 pub(super) mod iceberg_table_manager;
 mod iceberg_table_syncer;
 pub(super) mod index;
+pub(super) mod io_utils;
 pub(super) mod moonlink_catalog;
 pub(super) mod parquet_metadata_utils;
 pub(super) mod parquet_stats_utils;
